+++ crates/config/src/config/errors.rs
@@ -96,6 +96,9 @@ pub enum ConfigError {
         // #[diagnostic_source]
         #[source]
         error: ValidatorError,
+
+        #[help]
+        help: Option<String>,
     },
 }